# quarkus-astro-app/src/main/webui/api/openapi.yaml
...
openapi: 3.1.0
...
components:
  schemas:
    CategoryCreateDTO:
      type: object
      required:
      - name
      - colorCode
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 50
        colorCode:
          type: string
          pattern: "^#[0-9A-Fa-f]{6}$"
          minLength: 7
          maxLength: 7
    CategoryResponseDTO:
      type: object
      properties:
        id:
          $ref: "#/components/schemas/UUID"
        name:
          type: string
        colorCode:
          type: string
        isDefault:
          type: boolean
        createdAt:
          $ref: "#/components/schemas/LocalDateTime"
        taskCount:
          type: integer
          format: int64
    CategoryUpdateDTO:
      type: object
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 50
        colorCode:
          type: string
          pattern: "^#[0-9A-Fa-f]{6}$"
          minLength: 7
          maxLength: 7
...
    LocalDateTime:
      type: string
      format: date-time
      examples:
      - 2022-03-10T12:15:50
...
    UUID:
      type: string
      format: uuid
      pattern: "[a-fA-F0-9]{8}-[a-fA-F0-9]{4}-[a-fA-F0-9]{4}-[a-fA-F0-9]{4}-[a-fA-F0-9]{12}"
    UserDTO:
      type: object
      required:
      - username
      properties:
        username:
          type: string
          minLength: 1
          maxLength: 50
paths:
  /api/categories:
    get:
      summary: List user's categories
      description: Retrieve all categories for the authenticated user
      tags:
      - Categories
      responses:
        "200":
          description: List of categories
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CategoryResponseDTO"
        "401":
          description: Not authenticated
    post:
      summary: Create a new category
      description: Create a custom category with a unique name and color code
      tags:
      - Categories
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CategoryCreateDTO"
        required: true
      responses:
        "201":
          description: Category created successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CategoryResponseDTO"
        "400":
          description: Invalid input
        "401":
          description: Not authenticated
        "409":
          description: Category name already exists
  /api/categories/{id}:
    put:
      summary: Update a category
      description: Update a category's name and/or color code (PATCH semantics)
      tags:
      - Categories
      parameters:
      - description: Category UUID
        name: id
        required: true
        in: path
        schema:
          $ref: "#/components/schemas/UUID"
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CategoryUpdateDTO"
        required: true
      responses:
        "200":
          description: Category updated successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CategoryResponseDTO"
        "400":
          description: Invalid input
        "401":
          description: Not authenticated
        "404":
          description: Category not found
        "409":
          description: Category name already exists
    get:
      summary: Get category by ID
      description: Retrieve a specific category belonging to the authenticated user
      tags:
      - Categories
      parameters:
      - description: Category UUID
        name: id
        required: true
        in: path
        schema:
          $ref: "#/components/schemas/UUID"
      responses:
        "200":
          description: Category found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CategoryResponseDTO"
        "401":
          description: Not authenticated
        "404":
          description: Category not found
    delete:
      summary: Delete a category
      description: Delete a non-default category. Default categories cannot be deleted.
      tags:
      - Categories
      parameters:
      - description: Category UUID
        name: id
        required: true
        in: path
        schema:
          $ref: "#/components/schemas/UUID"
      responses:
        "204":
          description: Category deleted successfully
        "400":
          description: Cannot delete default category
        "401":
          description: Not authenticated
        "404":
          description: Category not found
  /api/session/login:
    post:
      summary: Login user
      description: Authenticates user with username only (demo mode). Creates session
        and seeds default categories on first login.
      tags:
      - Session
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserDTO"
        required: true
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserDTO"
        "400":
          description: Invalid request - username required
...
  /api/session/user:
    get:
      summary: Get current user
      description: Returns the currently authenticated user from session
      tags:
      - Session
      responses:
        "200":
          description: User is authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserDTO"
        "401":
          description: User is not authenticated
...
info:
  title: Task Manager API
  version: 1.0.0
  description: RESTful API for Task Manager application demonstrating Islands Architecture
  contact:
    name: Task Manager Team
servers:
- url: http://localhost:7171
  description: Auto generated value
- url: http://0.0.0.0:7171
  description: Auto generated value